Add a models header file.
